Add preliminary database down selection support
This commit adds functionality to down select on either models or
quilts. If any model(s) is/are provided to the script, all associated
quilts are returned/selected and any quilts passed to the script are
ignored.  Otherwise, if there is/are only quilt(s), their associated
boundaries are returned/selected. This may be of limited usefulness, but
I have added it for completeness. There are also corner cases or minutia
of this part of the script that I am not considering. However, in my
limited testing, it seems to do what I expect as of this commit.